feature(file buffer): Support in-memory file buffers
Enhance the HelloSign SDK to support file buffers instead of requiring
the use of temp-files. The support was added to include allowing the
caller to pass in the file name that is used by form.append( ) when
telling it that the Buffer should be treated as a file by giving it a
name.