[DEV-4086] Upgrade Travis CI test environment(s)
Previously, we were only running tests against the latest stable Node
version as well as iojs (which is outdated and became Node v4). Node 4,
as of 4/30/18 has now reached end of life.

We are now running tests against Node versions 9 and 10, as well as all
active LTS versions (non-maintenance) and the most recent stable version
of Node.