Resolve for responses with no body content
Successful 'cancel' requests will respond with a success status and no
content. `_responseHandler` didn't have a path for that, as it wasn't
checking for the case where the content-length header has a value of 0.

I'm not sure which cases will reach this path and have a content-length
of 1, so I've left that behaviour as it is so we don't break anything!

Resolves https://github.com/HelloFax/hellosign-nodejs-sdk/issues/77
Resolves https://github.com/HelloFax/hellosign-nodejs-sdk/issues/50
Resolves https://github.com/HelloFax/hellosign-nodejs-sdk/issues/21