Added ability to accept Buffers as files
We were really looking to use the Buffer capability of FormData, and the
only real issues we ran into were the ability to detect them, which
we've added, and the fact that the Buffers will have to have additional
properties assigned to them: .name and .contentType. If the user adds
these to the Buffer, once created with the contents of a file, then they
can simply put it in the files: [] argument to the function and it will
work just fine.